fix(clippy): resolve all 10 clippy warnings
- app: use Options struct init instead of field reassign (field_reassign_with_default)
- app: use checked_div for ops_per_sec and batch ops (manual_checked_ops)
- benchmark: remove unnecessary f64 cast and parens
- braces: use iterator with enumerate().skip() (needless_range_loop)
- main: auto-fix 4 collapsible_match patterns in key handlers
- tree: use next_back() instead of last() on DoubleEndedIterator

+++ picomatch-tui/src/ui/benchmark.rs
@@ -103,7 +103,7 @@ pub fn render_benchmark_tab(f: &mut Frame, app: &mut App, area: Rect) {
 
     // Left: Throughput & Gauge
     let ops_sec = app.benchmark_stats.ops_per_sec();
-    let gauge_ratio = ((ops_sec as f64 / 1_000_000.0).min(1.0)) as f64; // Max scaled at 1M ops/sec
+    let gauge_ratio = (ops_sec as f64 / 1_000_000.0).min(1.0); // Max scaled at 1M ops/sec
 
     let gauge_block = Block::default()
         .title(Span::styled(
